Separate into ui- & logic-focused files
Create cipher.c/.h for 'business logic',
ui.c/.h for the functions that get input,
main.c for the entry point.
Also minimalistic makefile and .gitignore.
