+++ main.c
@@ -1,0 +1,6 @@
+#include "ui.h"
+
+int main()
+{
+    return user_choice();
+}
