+++ makefile
@@ -1,0 +1,5 @@
+SRCS=$(wildcard *.c)
+
+OBJS=$(SRCS:.c=.o)
+
+main: $(OBJS)
